Section 7. Lecture 94
